Admin Roles, Dashboard, users and clean useless archives

+++ app/Http/Controllers/FileCategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\FileCategory;
 use App\Models\FileSubCategory;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class FileCategoryController extends Controller
 {
@@ -32,14 +33,17 @@ class FileCategoryController extends Controller
      */
     public function store(Request $request)
     {
-        $fileCat = new FileCategory();
+        if(Auth::user()->nivel_acesso_id == "1"){
+            $fileCat = new FileCategory();
+            $fileCat->name = trim($request->name);
+            $fileCat->iconMenu = trim($request->iconMenu);
+            $fileCat->href = trim($request->href);
+            $fileCat->save();
 
-        $fileCat->name = trim($request->name);
-        $fileCat->iconMenu = trim($request->iconMenu);
-        $fileCat->href = trim($request->href);
-        $fileCat->save();
-
-        return redirect()->route('manageFileCategories.index')->with('status', 'Categoria criada com sucesso!');
+            return redirect()->route('manageFileCategories.index')->with('status', 'Categoria criada com sucesso!');
+        } else{
+            return redirect()->route('home')->with('warning', 'Você não possui permissão para acessar esta página');
+        }
     }
 
     /**
@@ -50,8 +54,12 @@ class FileCategoryController extends Controller
      */
     public function show(FileCategory $FileCategory)
     {
-        $FileCategories = FileCategory::all();
-        return view('manageFileCategories.index', compact('FileCategories'));
+        if(Auth::user()->nivel_acesso_id == "1"){
+            $FileCategories = FileCategory::all();
+            return view('manageFileCategories.index', compact('FileCategories'));
+        } else{
+            return redirect()->route('home')->with('warning', 'Você não possui permissão para acessar esta página');
+        }
     }
 
     /**
@@ -62,9 +70,13 @@ class FileCategoryController extends Controller
      */
     public function edit(FileCategory $fileCategory)
     {
-        $cat = FileCategory::where(['id'=>$fileCategory->id])->first();
-        $subCategorias = FileSubCategory::where(['file_category_id'=>$fileCategory->id])->get();
-        return view('manageFileCategories.edit',compact('cat', 'subCategorias'));
+        if(Auth::user()->nivel_acesso_id == "1"){
+            $cat = FileCategory::where(['id'=>$fileCategory->id])->first();
+            $subCategorias = FileSubCategory::where(['file_category_id'=>$fileCategory->id])->get();
+            return view('manageFileCategories.edit',compact('cat', 'subCategorias'));
+        } else{
+            return redirect()->route('home')->with('warning', 'Você não possui permissão para acessar esta página');
+        }
     }
 
     /**
@@ -76,14 +88,17 @@ class FileCategoryController extends Controller
      */
     public function update(Request $request, FileCategory $fileCategory)
     {
+        if(Auth::user()->nivel_acesso_id == "1"){
+            FileCategory::where(['id'=>$fileCategory->id])->update([
+                'name' => trim($request->name),
+                'href' => trim($request->href),
+                'iconMenu' => trim($request->iconMenu),
+            ]);
 
-        FileCategory::where(['id'=>$fileCategory->id])->update([
-            'name' => trim($request->name),
-            'href' => trim($request->href),
-            'iconMenu' => trim($request->iconMenu),
-        ]);
-
-        return redirect()->route('manageFileCategories.index')->with('status', 'Categoria alterada com sucesso!');
+            return redirect()->route('manageFileCategories.index')->with('status', 'Categoria alterada com sucesso!');
+        } else{
+            return redirect()->route('home')->with('warning', 'Você não possui permissão para acessar esta página');
+        }
     }
 
     /**
@@ -94,6 +109,10 @@ class FileCategoryController extends Controller
      */
     public function destroy($id, FileCategory $FileCategory)
     {
-        $FileCategory->destroy($id);
+        if(Auth::user()->nivel_acesso_id == "1"){
+            $FileCategory->destroy($id);
+        } else{
+            return redirect()->route('home')->with('warning', 'Você não possui permissão para acessar esta página');
+        }
     }
 }
